refactor: ActivityHeatmapを共通コンポーネントに
